Append Last.fm data to albums that are missing it.
