+++ app/lib/parse.rb
@@ -4,7 +4,7 @@ class Parse
   def self.call(fan)
     case fan.provider
     when "last.fm"
-      Parsers::LastFm.call(fan)
+      Parsers::LastFm::UpdateFan.call(fan)
     when "spotify"
       Parsers::Spotify.call(fan)
     when NilClass
